--- HW7/Evgeniy_Nikulin/TadingSimulatorWebApi/Controllers/Clients/CountController.cs
@@ -1,21 +1,0 @@
-﻿namespace TadingSimulatorWebApi.Controllers.Clients
-{
-    using Microsoft.AspNetCore.Mvc;
-    using TradingSimulator.Core.Interfaces;
-
-    [ApiController]
-    [Route("clients/[controller]")]
-    public class CountController : ControllerBase
-    {
-        private readonly ITraderService traderService;
-        public CountController(ITraderService traderService) =>
-            this.traderService = traderService;
-
-        // GET: /clients/count
-        [HttpGet]
-        public string Get()
-        {
-            return traderService.GetTraderCount().ToString();
-        }
-    }
-}
