--- HW7/Evgeniy_Nikulin/TadingSimulatorWebApi/Controllers/Shares/RemoveController.cs
@@ -1,21 +1,0 @@
-﻿namespace TadingSimulatorWebApi.Controllers.Shares
-{
-    using Microsoft.AspNetCore.Mvc;
-    using TradingSimulator.Core.Interfaces;
-
-    [ApiController]
-    [Route("share/[controller]")]
-    public class RemoveController : ControllerBase
-    {
-        private readonly IShareService shareService;
-        public RemoveController(IShareService shareService) =>
-            this.shareService = shareService;
-
-        // POST: share/<controller>/
-        [HttpPost]
-        public void Post(int Id)
-        {
-            shareService.Remove(Id);
-        }
-    }
-}
